possibly fixing last issue however...
i'm having issues with the migrations.
the combatants table has no auto increment on the id column.
i have no idea how that happened and it probably happened during the other issues i had with the table.
i'll have to remove all migrations, destroy the db, and start fresh (or at least do this with the combatant table)