next and previous work
still undecided on this, but previous won't let you progress earlier than the first person in init in the round.
next however progresses the round. unsure if this is how i want this to work.

in addition, i think the current code is really poor. i'm sure there is a way to create one function to progress position that's shared by both calls. i also want to use jquery to avoid a page refresh (just adding/removing a class is what does it)